% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alcohol.example.R
\name{alcohol.example}
\alias{alcohol.example}
\title{\code{alcohol.example} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{alcohol.example} is a wrapper function for estimating a successful, safe merge.
}
