% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartools.example.R
\name{cartools.example}
\alias{cartools.example}
\title{\code{cartools.example} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{cartools.example} is a wrapper function for estimating a successful, safe merge.
}
