% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control1.model.R
\name{control1.model}
\alias{control1.model}
\title{\code{control1.model} is a wrapper for controlling speed to obtain a target speed.}
\usage{
control1.model()
}
\description{
\code{control1.model} is a wrapper for controlling speed to obtain a target speed.
}
