% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control2.model.R
\name{control2.model}
\alias{control2.model}
\title{\code{control2.model} is a wrapper for controlling speed to obtain a target speed.}
\usage{
control2.model(tdf1df11)
}
\arguments{
\item{tdf1df11}{speed and location data, data.frame}
}
\description{
\code{control2.model} is a wrapper for controlling speed to obtain a target speed.
}
