% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{H}
\alias{H}
\title{\code{H} is a wrapper function for plotting a ideal bottleneck merge.}
\usage{
H(t0,t1,u1,u2,x1,x2) uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{H} is a wrapper function for plotting a ideal bottleneck merge.
}
