% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske.model.R
\name{Helske.model}
\alias{Helske.model}
\title{\code{Helske.model} is a wrapper function for estimating a successful, safe merge. Mean speed = 53 mph = 78 fps.}
\usage{
Helske.model(usd)
}
\arguments{
\item{usd}{standard deviation of speed in mph, a number.}
}
\description{
\code{Helske.model} is a wrapper function for estimating a successful, safe merge. Mean speed = 53 mph = 78 fps.
}
