% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske10.model.R
\name{Helske10.model}
\alias{Helske10.model}
\title{\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom} See \code{Helske9.R}.}
\usage{
Helske10.model(tdf1df2)
}
\arguments{
\item{tdf1df10}{speed and location data, data.frame}
}
\description{
\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom} See \code{Helske9.R}.
}
