% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske11.model.R
\name{Helske11.model}
\alias{Helske11.model}
\title{\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom} See \code{Helske9.R}.}
\usage{
Helske11.model(tdf1df2)
}
\arguments{
\item{tdf1df11}{speed and location data, data.frame}
}
\description{
\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom} See \code{Helske9.R}.
}
