% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske12.model.R
\name{Helske12.model}
\alias{Helske12.model}
\title{\code{Helske12.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMtrend}.}
\usage{

}
\arguments{
\item{df}{speed and location data derived from \code{RingRoaddata}, data.frame}
}
\description{
\code{Helske12.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMtrend}.
}
