% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske2.model.R
\name{Helske2.model}
\alias{Helske2.model}
\title{\code{Helske2.model} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{Helske2.model}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\arguments{
\item{tdf1df2, }{time, speed and locations, a matrix}

\item{veh, }{a number}
}
\description{
\code{Helske2.model} is a wrapper function for estimating a successful, safe merge.
}
