% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske3.model.R
\name{Helske3.model}
\alias{Helske3.model}
\title{\code{Helske3.model} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{Helske3.model}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{Helske3.model} is a wrapper function for estimating a successful, safe merge.
}
