% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske7.model.R
\name{Helske7.model}
\alias{Helske7.model}
\title{\code{Helske7.model} is a wrapper function ring road experiment}
\usage{
uses \code{Helske4.model} is an extension of \code{Helske} models 1, 2 and 3
}
\arguments{
\item{usd}{standard deviation of speed in mph, a number.}

\item{veh, }{a number}
}
\description{
\code{Helske7.model} is a wrapper function ring road experiment
}
