% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske8.model.R
\name{Helske8.model}
\alias{Helske8.model}
\title{\code{Helske8.model} is a wrapper for learning how to fit \code{SSMarima} models.}
\usage{
Helske8.model(tdf1df2)
}
\arguments{
\item{tdf1df2}{speed and location data, data.frame}
}
\description{
\code{Helske8.model} is a wrapper for learning how to fit \code{SSMarima} models.
}
