% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske9.model.R
\name{Helske9.model}
\alias{Helske9.model}
\title{\code{Helske9.model} is a wrapper for learning how to fit \code{SSMarima} models.}
\usage{
Helske9.model(tdf1df2)
}
\arguments{
\item{tdf1df2}{speed and location data, data.frame}
}
\description{
\code{Helske9.model} is a wrapper for learning how to fit \code{SSMarima} models.
}
