% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linAcc.R
\name{linAcc}
\alias{linAcc}
\title{\code{linAcc} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{linAcc} is a wrapper function for estimating a successful, safe merge.
}
