% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergedlm.R
\name{mergedlm}
\alias{mergedlm}
\title{\code{mergedlm} is a wrapper function for plotting a ideal \code{zipper} and \code{side-by-side} bottleneck merges.}
\usage{
uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{mergedlm} is a wrapper function for plotting a ideal \code{zipper} and \code{side-by-side} bottleneck merges.
}
