% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEparam.R
\name{MLEparam}
\alias{MLEparam}
\title{\code{MLEparam} MLE.}
\usage{
MLEparam(tdf1fd2, veh, tend, frequency, m0, C0)
}
\arguments{
\item{tdf1fd2, }{observations from \code{cartools}, a matrix.}

\item{veh, }{vehicle number.}

\item{tend, }{end time in seconds}

\item{frequency, }{number of time-steps per second}

\item{m0, }{the expected value of the pre-sample state vector.}

\item{C0, }{the variance matrix of the pre-sample state vector.}
}
\description{
\code{MLEparam} MLE.
}
