% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RingRoaddata.R
\name{RingRoaddata}
\alias{RingRoaddata}
\title{\code{RingRoaddata} is a wrapper function used for testing \code{Helske} models.}
\usage{
RingRoaddata(usd, zsd)
}
\arguments{
\item{usd}{standard deviation of speed in mph, a number.}

\item{zsd}{standard deviation of measuring location, a number.}
}
\description{
\code{RingRoaddata} is a wrapper function used for testing \code{Helske} models.
}
