% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rw.R
\name{rw}
\alias{rw}
\title{\code{rw} Random walk model.}
\usage{
rw(tdf1df2)
}
\arguments{
\item{tdf1fd2, }{\code{cartools} data set, a matrix.}
}
\value{
Plot of the data, Kalman filters, smoother and forecasts.
}
\description{
\code{rw} Random walk model.
}
