% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwBuild.R
\name{rwBuild}
\alias{rwBuild}
\title{\code{rwBuild} estimate parameters of \code{rw} dlm model.}
\usage{
rwBuild(x)
}
\arguments{
\item{x, }{a vector.}
}
\description{
\code{rwBuild} estimate parameters of \code{rw} dlm model.
}
