% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwFilter.R
\name{rwFilter}
\alias{rwFilter}
\title{\code{rwFilter} Random walk model for filtering and smoothing.}
\usage{
rwFilter(tdf1df2)
}
\arguments{
\item{tdf1fd2, }{\code{cartools} data set, a matrix.}
}
\value{
Plots of the speed \code{u} data, Kalman filter and smoother.
}
\description{
\code{rwFilter} Random walk model for filtering and smoothing.
}
