% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwForecast.R
\name{rwForecast}
\alias{rwForecast}
\title{\code{rwForecast} Random walk model for forecasting.}
\usage{
rwForecast(tdf1df2, veh, dV, dW, m0, C0, tend, freqency, wv)
}
\arguments{
\item{veh, }{vehicle number.}

\item{dV, }{variance of the observation noise.}

\item{dW, }{diagonal elements of the variance matrix of the system noise.}

\item{m0, }{the expected value of the pre-sample state vector.}

\item{C0, }{the variance matrix of the pre-sample state vector.}

\item{tend, }{end time in seconds}

\item{wv}{signal noise ratio, a number}

\item{tdf1fd2, }{observations from \code{cartools}, a matrix.}

\item{frequency, }{number of time-steps per second}
}
\value{
A \code{ts} plot to study the effects of \code{dW} on the forecast.
}
\description{
All observations are in units of feet and seconds.
}
\examples{
rwForecast(tdf1df2, 1, 7.33, 7.33, 77.88, 77.88, 40, 8, 0.01)
}
