% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsmerge.R
\name{sbsmerge}
\alias{sbsmerge}
\title{\code{sbsmerge} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{xabparam}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{sbsmerge} is a wrapper function for estimating a successful, safe merge.
}
