\name{tdf1df2}
\alias{tdf1df2}
\docType{data}
\title{cartools time, speed and trajectory data}
\description{A data frame is generated with \code{cartools::zipper.simulate}}
\usage{data("tdf1df2")}
\format{
  A matrix for each vehicle in a merge.
  \describe{
    \item{\code{tseq}}{a numeric vector}
    \item{\code{u}}{a numeric vector}
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
