% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textbook.examples.R
\name{textbook.examples}
\alias{textbook.examples}
\title{\code{textbook.examples} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{textbook.examples}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{textbook.examples} is a wrapper function for estimating a successful, safe merge.
}
