% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xuBuild.R
\name{xuBuild}
\alias{xuBuild}
\title{\code{xuBuild} estimate parameters of location \code{x} and speed \code{u} dlm model.}
\usage{
xuBuild(x)
}
\arguments{
\item{x, }{a vector.}
}
\description{
\code{xuBuild} estimate parameters of location \code{x} and speed \code{u} dlm model.
}
