% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xuMLEparam.R
\name{xuMLEparam}
\alias{xuMLEparam}
\title{\code{xuMLEparam} MLE for \code{cartools} speed.}
\usage{
xuMLEparam(tdf1fd2, veh, tend, frequency)
}
\arguments{
\item{tdf1fd2, }{observations from \code{cartools}, a matrix.}

\item{veh, }{vehicle number.}

\item{tend, }{end time in seconds}

\item{frequency, }{observation per second, a number}
}
\description{
\code{xuMLEparam} MLE for \code{cartools} speed.
}
