% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske.model.R
\name{Helske.model}
\alias{Helske.model}
\title{\code{Helske.model} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{Helske.model}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\description{
\code{Helske.model} is a wrapper function for estimating a successful, safe merge.
}
