% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske2.model.R
\name{Helske2.model}
\alias{Helske2.model}
\title{\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom}.}
\usage{
Helske2.model(df)
}
\arguments{
\item{df}{speed and location data, data.frame}
}
\description{
\code{Helske2.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom}.
}
