% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske2.model.R
\name{Helske2.model}
\alias{Helske2.model}
\title{\code{Helske2.model} is a wrapper function for estimating a successful, safe merge. Mean speed = 53 mph = 78 fps.}
\usage{
Helske2.model(usd)
}
\arguments{
\item{usd}{standard deviation of speed in mph, a number.}
}
\description{
\code{Helske2.model} is a wrapper function for estimating a successful, safe merge. Mean speed = 53 mph = 78 fps.
}
