% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helske3.model.R
\name{Helske3.model}
\alias{Helske3.model}
\title{\code{Helske3.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom}.}
\usage{
uses \code{Helske3.model} is used to filter noisy speed data.
}
\arguments{
\item{df}{speed and location data, data.frame}
}
\description{
\code{Helske3.model} is a wrapper for estimating speed and location with Kalman filtering \code{SSMcustom}.
}
