% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.analysis.R
\name{merge.analysis}
\alias{merge.analysis}
\title{\code{merge.analysis} is a wrapper function for estimating a successful, safe merge.}
\usage{
uses \code{merge.analysis}, \code{xab} and \code{uab} from the \code{cartools} Package.
}
\arguments{
\item{tdf1df2, }{time, speed and locations, a matrix}
}
\description{
\code{merge.analysis} is a wrapper function for estimating a successful, safe merge.
}
